feat: introduce abstract base class for optical field definitions with concrete implementations for object height and paraxial image height
